Improve tests for custom book
